Use `DisabledComposerView` instead of no composer to add message
